src/requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-requests
robotframework-pabot
robotframework-faker
pyyaml
